authentication

+++ Personal-Blog-BE/personal-blogs-app/routers/personalBlogs.py
@@ -5,7 +5,8 @@ from pydantic import BaseModel, Field
 from datetime import datetime 
 from fastapi import APIRouter
 from database import SessionLocal
-from model import PersonalBlogs
+from models import PersonalBlogs
+from .auth import get_current_user
 
 
 router = APIRouter()
@@ -19,6 +20,7 @@ def get_db():
     finally:
         db.close()
 db_dependency = Annotated[Session , Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class BlogRequest(BaseModel):
     name:str = Field(min_length=3)
@@ -26,25 +28,37 @@ class BlogRequest(BaseModel):
     description:str
 
 @router.get("/blogs")
-async def read_all(db: db_dependency):
-    return db.query(PersonalBlogs).all()
+async def read_all(user : user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail='Authentication failed .')
+    return db.query(PersonalBlogs).filter(PersonalBlogs.owner_id == user.get('id')).all()
 
 @router.get("/blog/{blog_id}" , status_code = status.HTTP_200_OK)
-async def read_blogs(db:db_dependency , blog_id:int=Path(gt=0)):
-    blog_model = db.query(PersonalBlogs).filter(PersonalBlogs.id == blog_id).first()
+async def read_blogs(user : user_dependency, db:db_dependency , blog_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail='Authentication failed .')
+    blog_model = db.query(PersonalBlogs).filter(PersonalBlogs.id == blog_id).filter(PersonalBlogs.owner_id == user.get('id')).first()
     if blog_model is not None:
         return blog_model
     raise HTTPException(status_code = 404, detail = 'Blog not found')
 
 @router.post("/blog" , status_code = status.HTTP_201_CREATED)
-async def create_blog(db:db_dependency , blog_request:BlogRequest):
-    blog_model = PersonalBlogs(**blog_request.dict())
+async def create_blog(user : user_dependency, db:db_dependency , blog_request:BlogRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail='Authentication failed .')
+    blog_model = PersonalBlogs(**blog_request.dict() , owner_id = user.get('id'))
     db.add(blog_model)
     db.commit()
 
 @router.put("/blog/{blog_id}" , status_code = status.HTTP_204_NO_CONTENT)
-async def update_blog(db:db_dependency , blog_id:int , blog_request :BlogRequest):
-    blog_model = db.query(PersonalBlogs).filter(PersonalBlogs.id == blog_id).first()
+async def update_blog(user : user_dependency, db:db_dependency , blog_id:int , blog_request :BlogRequest):
+    blog_model = db.query(PersonalBlogs).filter(PersonalBlogs.id == blog_id).filter(PersonalBlogs.owner_id == user.get('id')).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail='Authentication failed .')
     if blog_model is None:
         raise HTTPException(status_code = 404, detail = 'Blog not found')
     blog_model.name = blog_request.name
@@ -55,8 +69,11 @@ async def update_blog(db:db_dependency , blog_id:int , blog_request :BlogRequest
     db.commit()
 
 @router.delete("/blog/{blog_id}" , status_code = status.HTTP_204_NO_CONTENT)
-async def delete_blog(db:db_dependency , blog_id:int):
-    blog_model = db.query(PersonalBlogs).filter(PersonalBlogs.id == blog_id).first()
+async def delete_blog(user : user_dependency, db:db_dependency , blog_id:int):
+    blog_model = db.query(PersonalBlogs).filter(PersonalBlogs.id == blog_id).filter(PersonalBlogs.owner_id == user.get('id')).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail='Authentication failed .')
     if blog_model is None:
         raise HTTPException(status_code = 404, detail = 'Blog not found')
     db.delete(blog_model)
